backlogs are now user exlusive

+++ routes/backlog.routes.js
@@ -5,12 +5,16 @@ const router = require("express").Router()
 
 // create a game listing - CREATE
 router.get ("/new",async (req, res)=>{
+    console.log(req.session.user)
     const allGames = await Backlog.find()
     res.render("backlogs/new.ejs", {allGames: allGames})
 })
 
 router.post('/', async (req,res)=>{
     try {
+        console.log(req.session.user)
+        req.body.user = req.session.user._id
+        
         console.log(req.body)
         if(req.body.status === "on"){
         req.body.status = true
@@ -25,7 +29,7 @@ router.post('/', async (req,res)=>{
 // shows the user their backlog - READ
 router.get("/", async (req,res)=>{
     try {
-        const userBacklog =  await Backlog.find()
+        const userBacklog =  await Backlog.find({user:req.session.user._id})
         console.log(userBacklog)
         // here i tell the page to render my 
         res.render('backlogs/userbacklog.ejs', {userBacklog: userBacklog})
@@ -56,8 +60,14 @@ router.get('/edit/:backlogId' ,async (req,res)=>{
 })
 router.put('/edit/:backlogId', async (req,res)=>{
     try {
+        let foundGame = Backlog.findById(req.params.backlogId)
+        // json stringify converts object id into strings
+         if ( req.session.user._id=== JSON.stringify(foundGame.user)){
+           return res.send('Error! this listing belongs to another user.')
+        }
         const allGames = await Backlog.find()
-        const foundGame = await Backlog.findByIdAndUpdate(req.params.backlogId, req.body)
+        foundGame = await Backlog.findByIdAndUpdate(req.params.backlogId, req.body)
+        
         res.redirect('/backlog')
     } catch (error) {
         console.log('failed to update game listing', error)
@@ -67,7 +77,13 @@ router.put('/edit/:backlogId', async (req,res)=>{
 // allows user to delete listings - DELETE
 router.delete('/:backlogId', async (req,res)=>{
     try {
-        const foundGame = await Backlog.findByIdAndDelete(req.params.backlogId)
+        let foundGame = await Backlog.findById(req.params.backlogId)
+        console.log("Session Id " + req.session.user._id)
+        console.log("found game id ", foundGame.user)
+        if ( req.session.user._id=== JSON.stringify(foundGame.user)){
+           return res.send('Error! this listing belongs to another user.')
+        }
+         foundGame = await Backlog.findByIdAndDelete(req.params.backlogId)
         res.redirect('/backlog')
     } catch (error) {
         console.log('failed to delete game listing', error)
